Use the p/z contribs for mag and phase of result instead of plot value
Before the change the value of the resulting phase got calculated from
the value of the point at the arrow head. The value therefore was
restricted to the range [-180°,180°]. This was confusing if there was
for example the contributions of -90°-90°-90° = -270° from three poles.
The result would display 90° and it's not immediately apparent why.

Now we calculate the result as a sum of the phase contributions.

Should fix #1