Fix (trail head) arrow annotations separating when zoomed in far
Up to now an arrow's tip and shaft would separate when its parent
plot is zoomed beyond 0.05 per grid delta.
(I regard this as a bug in Matlab's API.)

This fix changes the annotation's parent object to the figure itself.
As the figure's coordinates are relative in the range [0,1]x[0,1], we
need to perform a coordinate transform first. Use DS2NFU.m for that
purpose.

Note that the text arrow annotations of outlying p/z are still affected
by the separation bug, but this is acceptable as nobody probably will
zoom into the z-plot that far.

Reported this ill-behavior to Mathworks.
https://ch.mathworks.com/matlabcentral/answers/331372-potential-bug-in-annotation-arrow-with-parent-gca-when-zoomed-in-far

Fixes #2