Fix the 'zoom reset' issue that was introduced with v5.1.1
One can now double-click (with the zoom or pan tool) in a plot, then
it zooms out such that the whole data is visible.

The problem was with the documentation of zoom() and pan() that was
suggesting that one can only use a figure handle as the object to
operate on. zoom(hFig,__) and pan(hFig,__) will then operate on the
CurrentAxis, which sometimes is wrong.
Correct usage is zoom(hAx,__).

Fixes #4