Add 'pad_left.m', this function pads the shorter of two vectors with 0
Used as utility function for the addition of two vetors that
represent polynomials.

E.g. if we want to add [1 2 3], representing x^2 + 2x + 3,
and [1 -1], representing x - 1, then we have to pad the shorter one
with zeros first, i.e. [1 -1] becomes [0 1 -1]. Then we can simply
add the vectors, which gets us [1 3 2], or x^2 + 3x + 2.