Change the detour circles to go left instead of right
In the SISO case it doesn't matter if a detour goes around a pole or
zero to the left or right. But for MIMO systems one has to take a
left detour.

Suggested, for example, here:
http://www.tandfonline.com/doi/abs/10.1080/00207178108922569

Originally:
"On the Generalized Nyquist Stability Criterion," C.A. Desoer and Y.-T. Wang