Fix a bug where the plot's data isn't correctly updated after zoom reset
The problem is when one resets the zoom level with a double click.

'cb_after_zoom_or_pan' gets called twice in rapid succession and
'draw_run_continuous_animation' hasn't reset 'this.s_check_limits' yet,
so it doesn't know that it should update the display data a second time.