Change 'draw_update_result_textbox' to respect scaling factors
Terminology:
For a rational transfer function G(s)=a(s)/b(s) with polynomials a(s)
and b(s), call the ratio of the highest order coefficient of a(s) and
b(s) the 'scaling factor' of the transfer function.
E.g. if G(s)=(-2s+1)/(3s^2+5), then the scaling factor is -2/3.

Before the change, 'draw_update_result_textbox' would simply ignore the
effect of a scaling factor and instead show results for a transfer
function with the same poles and zeros but scaling factor = 1.

Now the method is aware of this factor and the results TextBox should
be consistent with the title of the z-plot (except for multiples of
360° in the phase).

Fixes #28