(non-functional) Integrate 'multroot' into 'polyratio.m'
Also: Variable rename of 'poles' and 'zeros' into 'p' and 'z' to
      avoid name collision with builtin function 'zeros'.

Multroot returns a 2D array containing the distinct root values in
the first column and their multiplicity in the second column.

This commit is non-functional because initialization of 'zeros_mult'
and 'poles_mult' is part of the next commit.