Crash 'multroot' when in infinite loop because of NaN value
There was a polynomial (which I unfortunately lost between
a Matlab restart) that caused 'multroot' to get stuck in an infinite
loop because 'bke(j)' somehow became NaN and thus the while loop
never breaking.

Added a test case which looks for this condition, should it ever
occur again, and then crash the function deliberately with an error
as security measurement.