Enable ANP to cope with constant transfer functions
Fixing this meant to change the behavior of the auto_zoom functions
that would crash or misbehave when no zeros and poles are present
whatsoever and the Nyquist plot becomes a point on the complex plane.

Not totally happy with the new xlim and ylim calculations (they can
produce non-optimal zoom reset values with the random test cases
sometimes, but this is a minor concern atm).

Closes #25